REF: Refactor file DetailHeroView yaitu memecah beberapa view menjadi components
